Clean bucket after bucket notification
