+++ tests/ctst/steps/notifications.ts
@@ -1,10 +1,11 @@
-import { Then, Given, When, setDefaultTimeout } from '@cucumber/cucumber';
+import { Then, Given, When, After, setDefaultTimeout } from '@cucumber/cucumber';
 import { strict as assert } from 'assert';
 import Zenko, {
     AWSVersionObject,
 } from '../world/Zenko';
 import { Constants, S3, Utils, KafkaHelper } from 'cli-testing';
 import { Message } from 'node-rdkafka';
+import { cleanZenkoBucket } from 'common/common';
 
 setDefaultTimeout(Constants.DEFAULT_TIMEOUT);
 
@@ -376,3 +377,10 @@ Then('i should {string} a notification for {string} event in destination {int}',
         const expected = receive === 'receive';
         assert.strictEqual(receivedNotification, expected);
     });
+
+After({ tags: '@BucketNotification' }, async function (this: Zenko) {
+    await cleanZenkoBucket(
+        this,
+        this.getSaved<string>('bucketName'),
+    );
+});
